Trigger draft storing from the lifecycle saga
